rename HardwareAbstractionInterface files
