--- Platformio/src/HardwareAbstractionInterface.cpp
@@ -1,1 +1,0 @@
-#include "HardwareAbstractionInterface.h"
